--- backend/src/main/java/com/f1veguys/sel/campaign/dto/CampaignRequest.java
@@ -1,13 +1,0 @@
-package com.f1veguys.sel.campaign.dto;
-
-
-import java.time.LocalDateTime;
-
-public record CampaignRequest(
-        String title,
-        int goalAmount,
-        LocalDateTime startDate,
-        LocalDateTime endDate,
-        boolean completed
-) {
-}
