--- backend/src/main/java/com/f1veguys/sel/campaignhistory/repository/CampaignHistoryRepository.java
@@ -1,7 +1,0 @@
-package com.f1veguys.sel.campaignhistory.repository;
-
-import com.f1veguys.sel.campaignhistory.domain.CampaignHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CampaignHistoryRepository extends JpaRepository<CampaignHistory, CampaignHistory.CampaignHistoryId> {
-}
