--- backend/src/main/java/com/f1veguys/sel/campaignhistory/service/CampaignHistoryService.java
@@ -1,7 +1,0 @@
-package com.f1veguys.sel.campaignhistory.service;
-
-import com.f1veguys.sel.campaignhistory.domain.CampaignHistory;
-
-public interface CampaignHistoryService {
-    CampaignHistory participateInCampaign(int campaignId, int userId, int contributionAmount);
-}
