+++ backend/src/main/java/com/f1veguys/sel/domain/campaign/dto/CampaignResponse.java
@@ -1,0 +1,25 @@
+package com.f1veguys.sel.domain.campaign.dto;
+
+import com.f1veguys.sel.domain.campaign.domain.Campaign;
+
+import java.time.LocalDateTime;
+
+public record CampaignResponse(
+        int id,
+        String title,
+        int goalAmount,
+        int nowAmount,
+        boolean completed,
+        LocalDateTime uploadDate
+) {
+    public CampaignResponse(Campaign campaign) {
+        this(
+                campaign.getId(),
+                campaign.getTitle(),
+                campaign.getGoalAmount(),
+                campaign.getNowAmount(),
+                campaign.isCompleted(),
+                campaign.getUploadDate()
+        );
+    }
+}
