+++ backend/src/main/java/com/f1veguys/sel/domain/campaign/service/CampaignService.java
@@ -1,0 +1,23 @@
+package com.f1veguys.sel.domain.campaign.service;
+
+import com.f1veguys.sel.domain.campaign.dto.CampaignResponse;
+
+import java.util.List;
+
+public interface CampaignService {
+//    CampaignResponse createCampaign(CampaignRequest request, List<MultipartFile> files) throws IOException;
+
+//    CampaignResponse updateCampaign(int id, CampaignRequest request);
+
+//    void deleteCampaign(int id);
+
+    CampaignResponse getCampaign(int campaignId);
+
+    List<CampaignResponse> getAllCampaigns();
+
+    List<CampaignResponse> getOngoingCampaigns();
+
+    List<CampaignResponse> getCompletedCampaigns();
+
+    void completeCampaign(int id);
+}
