+++ backend/src/main/java/com/f1veguys/sel/domain/campaignhistory/service/CampaignHistoryService.java
@@ -1,0 +1,7 @@
+package com.f1veguys.sel.domain.campaignhistory.service;
+
+import com.f1veguys.sel.domain.campaignhistory.domain.CampaignHistory;
+
+public interface CampaignHistoryService {
+    CampaignHistory participateInCampaign(int campaignId, int userId, int contributionAmount);
+}
