+++ backend/src/main/java/com/f1veguys/sel/domain/user/domain/User.java
@@ -1,0 +1,53 @@
+package com.f1veguys.sel.domain.user.domain;
+
+import com.f1veguys.sel.domain.campaignhistory.domain.CampaignHistory;
+import com.f1veguys.sel.domain.tree.domain.Tree;
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Entity
+@Table(name = "user")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Builder(toBuilder = true)
+@Getter
+@Setter
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "users_id")
+    private int id;
+
+    @Column(name = "email", nullable = false, unique = true)
+    private String email;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @Column(name = "password", nullable = false)
+    private String password;
+
+    @Column(name = "admin", nullable = false)
+    private boolean admin;
+
+    @Column(name = "api_id")
+    private String apiId;
+
+    @CreationTimestamp
+    @Column(name = "created_date", nullable = false, updatable = false)
+    private LocalDateTime createdDate;
+
+    @Column(name = "account_num", nullable = false, unique = true)
+    private String accountNum;
+
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<CampaignHistory> campaignHistories;
+
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<Tree> trees;
+}
