--- backend/src/main/java/com/f1veguys/sel/tree/domain/Tree.java
@@ -1,40 +1,0 @@
-package com.f1veguys.sel.tree.domain;
-
-import com.f1veguys.sel.user.domain.User;
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "tree_user")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder(toBuilder = true)
-@Getter
-@Setter
-public class Tree {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "tree_user_id")
-    private int id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    @JsonBackReference
-    private User user;
-
-    @Column(name = "count", nullable = false)
-    private int count;
-
-    @Column(name = "grown", nullable = false)
-    private boolean grown;
-
-    @CreationTimestamp
-    @Column(name = "created_date", nullable = false)
-    private LocalDateTime createdDate;
-
-}
