--- backend/src/main/java/com/f1veguys/sel/tree/service/TreeService.java
@@ -1,17 +1,0 @@
-package com.f1veguys.sel.tree.service;
-
-import com.f1veguys.sel.tree.domain.Tree;
-import com.f1veguys.sel.user.domain.User;
-
-public interface TreeService {
-
-    Tree getTree(int userId);
-
-    Tree startTree(User user);
-
-    void endTree(int id);
-
-    Tree waterTree(int id);
-
-    Tree getGift(int id);
-}
