--- backend/src/main/java/com/f1veguys/sel/user/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package com.f1veguys.sel.user.repository;
-
-
-import com.f1veguys.sel.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-}
